refactor: filter logic and modularize constants/utils
- Remove deprecated filter components and utility functions while consolidating filter logic into reusable utility and constants files.
- Improved readability, maintainability, and alignment with project structure by replacing scattered logic and centralizing shared resources.
