+++ app/(shop)/[slug]/[...category]/types.ts
@@ -1,0 +1,1 @@
+export type SortingOption = 'popular' | 'rating' | 'newest' | 'priceLow' | 'priceHigh';
